pyproject.toml
[tool.poetry]
name = "subtitles-transcript"
version = "0.1.0"
description = ""
authors = ["dongnt"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.1.0"}
loguru = "^0.7.3"
flask-cors = "^5.0.0"
python-dotenv = "^1.0.1"
m3u8 = "^6.0.0"
fireworks-ai = "^0.15.11"
requests = "^2.32.3"
aiohttp = "^3.11.11"
aiofiles = "^24.1.0"
langdetect = "^1.0.9"
pysrt = "^1.1.2"
openai = "^1.59.6"
pydub = "^0.25.1"
uwsgi = "^2.0.28"
yt-dlp = "^2024.12.23"
google-generativeai = "^0.8.4"
clean-text = "^0.6.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
